refactor: move constants.ts to constants/api.ts
