+++ dashboard/src/constants/api.ts
@@ -1,8 +1,3 @@
-export const RESPONSE_STATUS = {
-    OK: 'OK',
-    FAILED: 'FAILED',
-};
-
 export const API_PATH_PREFIX = import.meta.env.VITE_API_PATH_PREFIX;
 
 export const END_POINT = {
@@ -10,4 +5,9 @@ export const END_POINT = {
     IS_SIGNED: 'auth/is-sign-in',
     DEVICES: 'me/devices',
     DEVICE: 'me/devices/:id',
+};
+
+export const RESPONSE_STATUS = {
+    OK: 'OK',
+    FAILED: 'FAILED',
 };
